requirements.txt
aniso8601==9.0.1
click==7.1.2
colorama==0.4.4
coverage==5.5
Faker==14.0.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
PyJWT==2.1.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==2.0.1
gunicorn==20.1.0
numpy==1.23.3
pydub==0.25.1
ffmpeg==1.4
rq==1.1.0 
celery==5.0.5
python-decouple==3.6
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
psycopg2-binary==2.8.6
pytest-cov==4.0.0
python-dateutil==2.8.2